Day of week snappers shouldn't touch the time
This commit will change the behaviour of the day of week snappers in the
AST so that they just add or substract days, without moving to the start
or to the end of a day.

This would make the snapper perform two things, and it's probably better
to use two consecutive snappers to modify a date and a time at the same
time.