fix: error message on import due to wrong main path
Upon importing the module in a project or even running `require(".")` in
a Node.js REPL inside the working directory, the following warning is
visible:

   (node:7) [DEP0128] DeprecationWarning: Invalid 'main' field in
     '/app/node_modules/datetoken/package.json' of './lib/index.js'.
     Please either fix that or report it to the module author

Given that the lib/ directory does not exist in the repository, this
commit will update the package.json so that it uses the proper path. The
package uploaded to NPM is also having the index.js and index.d.ts file
in the repository root.